Fetch rule in controller, not testing service
